Only use mysql creation and dropping if specified in testjour.yml

+++ lib/testjour/commands/local_run.rb
@@ -14,16 +14,28 @@ module Commands
       daemonize
       Testjour.logger.info "Starting local:run"
       
-      mysql = MysqlDatabaseSetup.new
-      mysql.create_database
-      ENV["TESTJOUR_DB"] = mysql.runner_database_name
+      mysql = false
       
-      silence_stream(STDOUT) do
-        system schema_load_command(mysql.runner_database_name)
+      if File.exist?("testjour.yml")
+        testjour_yml = File.read("testjour.yml")
+        
+        if testjour_yml.include?("--mysql")
+          mysql = true
+        end
       end
       
-      at_exit do
-        mysql.drop_database
+      if mysql
+        mysql = MysqlDatabaseSetup.new
+        mysql.create_database
+        ENV["TESTJOUR_DB"] = mysql.runner_database_name
+      
+        silence_stream(STDOUT) do
+          system schema_load_command(mysql.runner_database_name)
+        end
+      
+        at_exit do
+          mysql.drop_database
+        end
       end
       
       initialize_cucumber
